refactor: separate short URL redirect logic into a dedicated component for improved maintainability and clarity
